create story

create story

create_story(
  name,
  description,
  epic_id,
  owner_ids,
  project_id = NULL,
  workflow_state_id,
  group_id,
  archived = c(FALSE, TRUE),
  story_type = c("feature", "bug", "chore"),
  created_at = Sys.time(),
  estimate = NULL,
  iteration_id = NULL,
  deadline = NULL,
  labels = NULL,
  config = list(),
  sc_token = get_token()
)

Arguments

name

Character (Required). The Story’s name.

description

Character (Required). The description of the story.

epic_id

Numeric (Required). The ID of the epic the story belongs to. To get the Epic ID, see get_all("epics")$id

owner_ids

List of Strings (Required). An list of UUIDs of the owners of this story.To get the Owner UUID, see get_all("members")$id

project_id

Numeric (Required). The ID of the project the story belongs to. To get the Project ID, see get_all("projects")$id. Note projects must be enabled for your team

workflow_state_id

Numeric (Required). The ID of the workflow state the story will be in. To get the Workflow State ID, see get_all("workflows")$states$id

group_id

Character (Required). The ID of the group to associate with this story. To get the Group ID, see get_all("groups")$id

archived

Character (Optional). Controls the story’s archived state. Defaults to FALSE

story_type

Character (Optional). The type of story (feature, bug, chore). Defaults to feature.

created_at

POSIXct (optional). The time/date the Story was created. Defaults to Sys.time().

estimate

Numeric (Optional). The numeric point estimate of the story. Defaults to NULL, which means unestimated.

iteration_id

Numeric (Optional). The ID of the iteration the story belongs to. To get the iteration_id, see get_all("iterations")$id

deadline

POSIXct (Optional). The due date of the story.

labels

List of Labels (Optional). A list of labels attached to the story.

config

(optional) additional configuration to add to header

sc_token

shortcut API token. Defaults to get_token()

Examples

if (FALSE) {
create_story(
  name = "Test Story",
  epic_id = 123,
  owner_ids = list("12345678-9012-3456-7890-123456789012"),
  workflow_state_id = 123,
  project_id = NULL,
  group_id = "12345678-9012-3456-7890-123456789012",
  archived = FALSE,
  story_type = "feature",
  description = "Test Description"
)

label_story <- list(
  list(
      color = '#49a940',
      name = 'in sprint'
    )
    , list(
      color = '#cc5856',
      name = 'new request'
    )
)

create_story(
  name = "Test Story",
  epic_id = 123,
  owner_ids = list("12345678-9012-3456-7890-123456789012"),
  workflow_state_id = 123,
  project_id = 123,
  group_id = "12345678-9012-3456-7890-123456789012",
  archived = FALSE,
  story_type = "feature",
  description = "Test Description",
  labels = label_story
)
}